Add files via upload
Objective: Forest vs. Building Classification Using ANN In the dataset, there are two files: the forest and the building file. Each file consists of forest and building images. We are going to convert the image file to an array, which consists of the pixels of the images. Then we are going to create an ANN architecture and predict on test data images.